% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclideanDistance.R
\name{euclideanDistance}
\alias{euclideanDistance}
\title{Euclidean distance function}
\usage{
euclideanDistance(x, y)
}
\arguments{
\item{x}{x-axis entry}

\item{y}{y-axis entry}
}
\value{
The Euclidean distance from (x,y) to (0,0)
}
\description{
This is my first function and it creates the euclidean distance to the origin
in two dimensions
}
\note{
This function is a bit boring but that is ok.
}
\examples{
euclideanDistance(1, 2)
}
