% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximumDistance.R
\name{maximumDistance}
\alias{maximumDistance}
\title{Maximum distance function}
\usage{
maximumDistance(x, y)
}
\arguments{
\item{x}{x-axis entry}

\item{y}{y-axis entry}
}
\value{
The Manhattan distance from (x,y) to (0,0)
}
\description{
This is my first function and it creates the Maximum distance to the origin
in two dimensions
}
\note{
This function is a bit boring but that is ok.
}
\examples{
maximumDistance(1, 2)
}
