% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notAnotherRPackage-package.R
\docType{package}
\name{notAnotherRPackage-package}
\alias{notAnotherRPackage}
\alias{notAnotherRPackage-package}
\title{notAnotherRPackage: The Package that just does some stuff}
\description{
With this package, you will do some really interesting things.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jokrisam.github.io/notAnotherRPackage/}
}

}
\author{
\strong{Maintainer}: Johannes Krisam \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
